Add solution for problem `F - Putting Boxes Together` Link: https://codeforces.com/contest/1030/problem/F Tags: data structures, *2500 Ref: https://codeforces.com/contest/1030/submission/109036401

+++ codeforces/1030/F.cpp
@@ -239,7 +239,7 @@ void solve(int) {
     
     int n, q;
     cin >> n >> q;
-    vector<int> a(n + 1);
+    vector<ll> a(n + 1);
     vector<ll> w(n + 1);
     Fenwick<ll> f(n + 2);
     Fenwick<mint> fmod(n + 2);
